#2 local testing mode with env variable
