+++ .gitignore
@@ -2,4 +2,3 @@ coverage
 npm-debug.log
 node_modules
 creds
-local.js
